# backend/src/main/resources/openapi-docs.yaml
openapi: 3.0.1
info:
  title: API Ristorino
  description: |
    Portal gastronómico para descubrir y reservar en restaurantes de Córdoba.
    
    ## Autenticación
    
    La API usa autenticación JWT (JSON Web Tokens) stateless. Para acceder a endpoints protegidos:
    
    1. Registrar nuevo cliente con `POST /api/clientes/register` o hacer login con `POST /api/clientes/login`
    2. Obtener el token JWT de la respuesta
    3. Incluir el token en el header de las requests: `Authorization: Bearer <token>`
    
    ## Logout
    
    Con JWT stateless no hay sesión del lado del servidor. El logout se realiza del lado del cliente:
    - Eliminar el token JWT del almacenamiento local (localStorage, cookies, etc.)
    - No hay endpoint de logout en el servidor ya que el token es independiente del servidor
  version: 1.0.0
  contact:
    name: Equipo DAS - UBP
    email: soporte@ristorino.com
servers:
  - url: http://localhost:8080
    description: Servidor de Desarrollo
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido desde POST /api/clientes/register o POST /api/clientes/login
...
security:
  - BearerAuth: []
paths:
...
  /api/reservas:
    get:
      tags:
        - Reservas
      summary: Obtener todas las reservas
      description: Retorna la lista de reservas (requiere autenticación)
      responses:
        '200':
          description: Lista de reservas obtenida exitosamente
        '401':
          description: No autenticado
    post:
      tags:
        - Reservas
      summary: Crear una reserva
      description: Crea una nueva reserva (requiere autenticación)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Reserva creada exitosamente
        '400':
          description: Datos inválidos
        '401':
          description: No autenticado
  /api/reservas/{id}:
    get:
      tags:
        - Reservas
      summary: Obtener reserva por ID
      description: Retorna los detalles de una reserva específica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reserva encontrada
        '401':
          description: No autenticado
        '404':
          description: Reserva no encontrada
    put:
      tags:
        - Reservas
      summary: Actualizar una reserva
      description: Actualiza los datos de una reserva existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reserva actualizada exitosamente
        '401':
          description: No autenticado
        '404':
          description: Reserva no encontrada
    delete:
      tags:
        - Reservas
      summary: Eliminar una reserva
      description: Elimina una reserva existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Reserva eliminada exitosamente
        '401':
          description: No autenticado
        '404':
          description: Reserva no encontrada
  /api/reservas/{id}/estado:
    put:
      tags:
        - Reservas
      summary: Cambiar estado de una reserva
      description: Actualiza el estado de una reserva
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estado actualizado exitosamente
        '401':
          description: No autenticado
        '404':
          description: Reserva no encontrada
...